Move and rename MultiValueSearchTree encoding test
